electricity2030: split core files
