+++ src/climatevision/generator/electricity2030/commune/electricity2030_commune.py
@@ -16,21 +16,20 @@ from ...transport2030.t30 import T30
 from ...waste2030 import WasteLines
 
 from ..e30 import E30
-from ..core.electricity2030_core import (
-    FossilFuelsProduction,
-    EColVars2030,
-    EnergyDemand,
-    Energy,
-    calc_production_renewable_geothermal,
-    calc_stop_production_by_fossil_fuels,
-    calc_production_local_pv_roof,
-    calc_production_local_pv_facade,
+from ..core.energy import Energy, EnergyDemand
+from ..core.fossil_fuels_production import FossilFuelsProduction
+from ..core.e_col_vars_2030 import EColVars2030
+from ..core.geothermal import calc_production_renewable_geothermal
+from ..core.fossil_fuels import calc_stop_production_by_fossil_fuels
+from ..core.pv import (
     calc_production_local_pv_agri,
+    calc_production_local_pv_facade,
     calc_production_local_pv_park,
-    calc_production_local_wind_onshore,
-    calc_renew_wind_offshore,
-    calc_production_local_hydro,
+    calc_production_local_pv_roof,
 )
+from ..core.wind import calc_production_local_wind_onshore, calc_renew_wind_offshore
+from ..core.hydro import calc_production_local_hydro
+
 
 from .energy_demand.calc_energy_demand import calc_energy_demand
 from .energy_demand.calc_production_local_biomass_stage2 import (
