+++ src/climatevision/generator/electricity2030/core/e_col_vars_2030.py
@@ -1,0 +1,54 @@
+# pyright: strict
+
+from dataclasses import dataclass
+
+from .energy import EnergyDemandWithCostFuel
+
+# Definition der relevanten Spaltennamen für den Sektor E
+@dataclass(kw_only=True)
+class EColVars2030(EnergyDemandWithCostFuel):
+    pet_sites: float = None  # type: ignore
+    energy_installable: float = None  # type: ignore
+    cost_fuel_per_MWh: float = None  # type: ignore
+    mro_per_MWh: float = None  # type: ignore
+    mro: float = None  # type: ignore
+    CO2e_combustion_based_per_MWh: float = None  # type: ignore
+    CO2e_combustion_based: float = None  # type: ignore
+    cost_certificate_per_MWh: float = None  # type: ignore
+    cost_certificate: float = None  # type: ignore
+    cost_climate_saved: float = None  # type: ignore
+    cost_mro: float = None  # type: ignore
+    cost_mro_pa_com: float = None  # type: ignore
+    CO2e_total: float = None  # type: ignore
+    CO2e_total_2021_estimated: float = None  # type: ignore
+    demand_electricity: float = None  # type: ignore
+    demand_emplo: float = None  # type: ignore
+    demand_emplo_com: float = None  # type: ignore
+    power_installed: float = None  # type: ignore
+    power_to_be_installed_pct: float = None  # type: ignore
+    power_to_be_installed: float = None  # type: ignore
+    power_installable: float = None  # type: ignore
+    area_ha_available: float = None  # type: ignore
+    area_ha_available_pct_of_action: float = None  # type: ignore
+    ratio_power_to_area_ha: float = None  # type: ignore
+    cost_mro_pa: float = None  # type: ignore
+    change_CO2e_t: float = None  # type: ignore
+    change_CO2e_pct: float = None  # type: ignore
+    change_cost_energy: float = None  # type: ignore
+    change_cost_mro: float = None  # type: ignore
+    invest: float = None  # type: ignore
+    invest_pa: float = None  # type: ignore
+    invest_com: float = None  # type: ignore
+    invest_pa_com: float = None  # type: ignore
+    invest_outside: float = None  # type: ignore
+    invest_pa_outside: float = None  # type: ignore
+    invest_per_x: float = None  # type: ignore
+    pct_of_wage: float = None  # type: ignore
+    pct_x: float = None  # type: ignore
+    ratio_wage_to_emplo: float = None  # type: ignore
+    cost_wage: float = None  # type: ignore
+    cost_mro_per_MWh: float = None  # type: ignore
+    emplo_existing: float = None  # type: ignore
+    demand_emplo_new: float = None  # type: ignore
+    full_load_hour: float = None  # type: ignore
+    lifecycle: float = None  # type: ignore
