+++ src/climatevision/generator/electricity2030/core/energy.py
@@ -1,0 +1,20 @@
+# pyright: strict
+
+from dataclasses import dataclass
+
+
+@dataclass(kw_only=True)
+class Energy:
+    energy: float = None  # type: ignore
+
+
+@dataclass(kw_only=True)
+class EnergyDemand(Energy):
+    change_energy_MWh: float = None  # type: ignore
+    change_energy_pct: float = None  # type: ignore
+
+
+@dataclass(kw_only=True)
+class EnergyDemandWithCostFuel(EnergyDemand):
+    cost_fuel_per_MWh: float = None  # type: ignore
+    cost_fuel: float = None  # type: ignore
