+++ src/climatevision/generator/electricity2030/core/renewable_geothermal_production.py
@@ -1,0 +1,38 @@
+# pyright: strict
+
+from dataclasses import dataclass
+
+from .energy import EnergyDemand
+
+
+@dataclass(kw_only=True)
+class RenewableGeothermalProduction(EnergyDemand):
+    """Energy production using geothermal."""
+
+    energy_installable: float = None  # type: ignore
+    CO2e_combustion_based_per_MWh: float = None  # type: ignore
+    CO2e_combustion_based: float = None  # type: ignore
+    cost_climate_saved: float = None  # type: ignore
+    cost_mro: float = None  # type: ignore
+    CO2e_total: float = None  # type: ignore
+    demand_emplo: float = None  # type: ignore
+    power_installed: float = None  # type: ignore
+    power_to_be_installed_pct: float = None  # type: ignore
+    power_to_be_installed: float = None  # type: ignore
+    power_installable: float = None  # type: ignore
+    change_CO2e_t: float = None  # type: ignore
+    change_CO2e_pct: float = None  # type: ignore
+    change_cost_mro: float = None  # type: ignore
+    invest: float = None  # type: ignore
+    invest_pa: float = None  # type: ignore
+    invest_outside: float = None  # type: ignore
+    invest_pa_outside: float = None  # type: ignore
+    invest_per_x: float = None  # type: ignore
+    pct_of_wage: float = None  # type: ignore
+    pct_x: float = None  # type: ignore
+    ratio_wage_to_emplo: float = None  # type: ignore
+    cost_wage: float = None  # type: ignore
+    cost_mro_per_MWh: float = None  # type: ignore
+    emplo_existing: float = None  # type: ignore
+    demand_emplo_new: float = None  # type: ignore
+    full_load_hour: float = None  # type: ignore
